requirements.txt
absl-py==0.7.0
astor==0.7.1
cycler==0.10.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
imageio==2.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mock==3.0.5
numpy==1.16.1
pandas==0.24.1
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tflearn==0.3.2
Werkzeug==0.14.1
